Harden decoders/EXIF and fix correctness issues from full code review
Second comprehensive review pass (full line-by-line read of all 177
first-party files). Applies 47 verified fixes across 36 files; both
Release|x64 and Debug|x64 build clean.

Security (untrusted-file paths):
- dcraw ppm_thumb: fix heap over-read (copy loop indexed the source with
  the padded RGBWIDTH stride while the buffer used the unpadded stride)
  and add the missing malloc NULL check; same stride fix in ppm16_thumb
- EXIFReader: reject APP1 blocks too small for the TIFF header/IFD0 offset
  (nApp1Size < 18) and bound the thumbnail SOF segment before reading
  pSOF[5..8]
- Widen remaining 32-bit size math to size_t: RAWWrapper stride*height,
  CXMMImage::GetMemSize, ConvertToDIBRGBA, and six ImageLoadThread
  alpha-blend loop counters; add MAX_IMAGE_DIMENSION/PIXELS guards to
  AVIFWrapper
- JXL: free the file buffer on ReadFile failure / decoder-setup failure
  (buffer ownership now adopted at decoder entry, kept double-free-safe)
- FileList: NUL-terminate the slide-show probe buffer before strstr/strchr
- JPEGImage: guard against NULL pixels when 1-channel BGRA conversion fails
- Clipboard/PSD: GlobalLock/SetClipboardData NULL and leak checks, reject
  negative DIB width, cap the PSD ICC-profile resource size

Correctness / robustness:
- SettingsProvider: ReadIniString now honors its fileName arg (the global
  INI fallback was reading the user INI); m_zoomPauseFactor is now double
- KeyMap: compare key names with _tcscmp, not pointer == (StringPooling is
  disabled, so the +/- shortcuts rendered as "Plus"/"Minus")
- MainDlg: fix always-true transition frame-time test; initialize
  m_pDirectoryWatcher; free m_pImageProcParams2; validate WM_COPYDATA size
- BatchCopyDlg/ManageOpenWithDlg: fix edit-buffer off-by-one stack writes
- HelpDisplay: honor the per-line info color; NULL-guard Key on arena
  overflow
- FileExtensionsRegistry: fix inverted ACE-type precedence test;
  NUL-terminate registry string reads
- HEIF: check dimensions before the full decode
- JPEGView, UserCommand, NLS, DesktopWallpaper, PaintMemDCMgr,
  HashCompareLPCTSTR, CropCtl, GUIControls, Tooltip, ImageProcessingPanel,
  SaveImage, DirectoryWatcher, Helpers/MaxImageDef: uninitialized reads,
  a registry-key/HGLOBAL leak, off-by-one hash, dead state removal, an
  out-of-bounds command-line read, and a locale size-argument fix

+++ src/JPEGView/RAWWrapper.cpp
@@ -40,7 +40,7 @@ CJPEGImage* RawReader::ReadImage(LPCTSTR strFileName, bool& bOutOfMemory, bool b
 
 		int stride = Helpers::DoPadding(width * colors, 4);
 		
-		pPixelData = new(std::nothrow) unsigned char[stride * height];
+		pPixelData = new(std::nothrow) unsigned char[(size_t)stride * height];
 		if (pPixelData == NULL) {
 			bOutOfMemory = true;
 			return NULL;
